refactor: createdEmojiList from nullable to non-nullable, initial state & error in createEmoji -> createEmojiList is emptyList

+++ android/app/src/main/java/com/goliath/emojihub/data_sources/local/X3dDataSource.kt
@@ -32,7 +32,7 @@ interface X3dDataSource {
         inferenceResults: List<X3dInferenceResult>,
         classNameFilePath: String,
         classUnicodeFilePath: String
-    ): List<CreatedEmoji>?
+    ): List<CreatedEmoji>
 }
 
 @Singleton
@@ -184,14 +184,14 @@ class X3dDataSourceImpl @Inject constructor(
         inferenceResults: List<X3dInferenceResult>,
         classNameFilePath: String,
         classUnicodeFilePath: String
-    ): List<CreatedEmoji>? {
+    ): List<CreatedEmoji> {
         val classNameJSONObject = JSONObject(File(classNameFilePath).readText())
         val classUnicodeJSONObject = JSONObject(File(classUnicodeFilePath).readText())
 
         val createdEmojiList = mutableListOf<CreatedEmoji>()
         for (result in inferenceResults) {
-            val className = classNameJSONObject.getString(result.scoreIdx.toString()) ?: return null
-            val classUnicode = classUnicodeJSONObject.getString(className) ?: return null
+            val className = classNameJSONObject.getString(result.scoreIdx.toString()) ?: return emptyList()
+            val classUnicode = classUnicodeJSONObject.getString(className) ?: return emptyList()
             createdEmojiList.add(CreatedEmoji(className, classUnicode))
         }
         return createdEmojiList
@@ -200,9 +200,9 @@ class X3dDataSourceImpl @Inject constructor(
     fun assetFilePath(assetName: String): String {
         val file = File(context.filesDir, assetName.split("/").last())
         // FIXME: assetFilePath로 호출하고자 하는 파일에 변경사항(개발자 관점)이 생길 시 반영할 수 없음
-//        if (file.exists() && file.length() > 0) {
-//            return file.absolutePath
-//        }
+        if (file.exists() && file.length() > 0) {
+            return file.absolutePath
+        }
         context.assets.open(assetName).use { inputStream ->
             FileOutputStream(file).use { outputStream ->
                 val buffer = ByteArray(4 * 1024)
